+++ tests/test_serde.py
@@ -1,6 +1,4 @@
 import loopgpt
-import builtins
-import os
 
 from loopgpt.tools import Browser
 
@@ -15,18 +13,3 @@ def test_serde_browser():
     cfg = browser.config()
     browser = Browser.from_config(cfg)
     assert browser.browser_type == "firefox"
-
-def test_add_key_prompt(monkeypatch):
-    # Set up mock user input
-    responses = iter(["y", "my-api-key"])
-    monkeypatch.setattr(builtins, "input", lambda _: next(responses))
-
-    from loopgpt.utils.openai_key import check_openai_key
-    # Run the function
-    check_openai_key()
-
-    # Assert that the .env file was created with the correct key value
-    with open(".env", "r") as f:
-        assert f.read().strip() == 'OPENAI_API_KEY = "my-api-key"'
-    
-    os.remove(".env")
